pyproject.toml
[project]
name = "moonandsun-calendar"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "astral>=3.2",
    "ephem>=4.1.6",
    "google-api-python-client>=2.149.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
    "pytz>=2024.2",
    "ruff>=0.7.0",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["I"]
